refactor: extract DisconnectNotice into separate class
